refactor(platform): migrate the macOS platform module to objc2
`MacOSPlatform` now holds a typed `Retained<NSApplication>` instead of a raw
`id`: `sharedApplication` is reached through `MainThreadMarker` (which both
enforces the main-thread contract and produces the token the typed API needs,
so the old `debug_assert_appkit_main_thread` no longer needs to run first in
`with_config`), `setActivationPolicy`/`activateIgnoringOtherApps`/`run`/
`terminate` are typed calls, and `keyWindow` answers `Option<Retained<NSWindow>>`
rather than being nil-checked by hand. `app_path` uses typed
`NSBundle::mainBundle`/`bundlePath`, and the debug thread check is
`MainThreadMarker::new()` — the same `+[NSThread isMainThread]` question the
old `msg_send!` asked, in objc2's typed form.

`display.rs`'s `refresh_period_for_screen` boundary is unchanged (raw pointer
in, typed borrow inside) because `window.rs` is still on the older stack.

Verified on this Mac: `just macos-close-path` PASS, `just macos-frame-pump`
PASS, `cargo test -p flui-platform --lib` 189 passed, `just ci` green.

+++ crates/flui-platform/src/platforms/macos/display.rs
@@ -12,7 +12,7 @@ use std::sync::Arc;
 use objc2::MainThreadMarker;
 use objc2::rc::Retained;
 use objc2_app_kit::NSScreen;
-use objc2_foundation::{NSArray, NSDictionary, NSNumber, NSRect, NSString};
+use objc2_foundation::{NSArray, NSDictionary, NSNumber, NSString};
 
 use flui_types::geometry::{Bounds, DevicePixels, Point, Size};
 
@@ -87,7 +87,7 @@ fn display_id_of(screen: &NSScreen) -> u64 {
     else {
         return 0;
     };
-    value.unsignedLongLongValue() as u64
+    value.unsignedLongLongValue()
 }
 
 impl PlatformDisplay for MacOSDisplay {
